move Semaphore utility class into its own file
