+++ packages/common/concurrency.js
@@ -6,7 +6,6 @@ const https = require('follow-redirects').https;
 const url = require('url');
 const pLimit = require('p-limit');
 const log = require('./log');
-const ResourcesLockedError = require('./errors').ResourcesLockedError;
 
 /**
  * Wrap a function to limit how many instances can be run in parallel
@@ -79,76 +78,6 @@ const promiseUrl = (urlstr) =>
     }).on('error', reject);
   });
 
-class Semaphore {
-  constructor(docClient, tableName) {
-    this.docClient = docClient;
-    this.tableName = tableName;
-  }
-
-  up(key) {
-    return this.add(key, 1);
-  }
-
-  down(key) {
-    return this.add(key, -1);
-  }
-
-  async checkout(key, count, max, fn) {
-    let result = null;
-    log.info(`Incrementing ${key} by ${count}`);
-    try {
-      await this.add(key, count, max);
-    } catch (e) {
-      if (e.message === 'The conditional request failed') {
-        throw new ResourcesLockedError(`Could not increment ${key} by ${count}`);
-      }
-      log.error(e.message, e.stack);
-      throw e;
-    }
-    try {
-      result = await fn();
-    } finally {
-      log.info(`Decrementing ${key} by ${count}`);
-      await this.add(key, -count);
-    }
-    return result;
-  }
-
-  async add(key, count, max = 0) {
-    try {
-      const params = {
-        TableName: this.tableName,
-        Item: {
-          key: key,
-          semvalue: 0
-        },
-        ConditionExpression: '#key <> :key',
-        ExpressionAttributeNames: { '#key': 'key' },
-        ExpressionAttributeValues: { ':key': key }
-      };
-      await this.docClient.put(params).promise();
-    } catch (e) {
-      if (e.code !== 'ConditionalCheckFailedException') {
-        throw e;
-      }
-    }
-
-    const updateParams = {
-      TableName: this.tableName,
-      Key: { key: key },
-      UpdateExpression: 'set semvalue = semvalue + :val',
-      ExpressionAttributeValues: { ':val': count },
-      ReturnValues: 'UPDATED_NEW'
-    };
-
-    if (count > 0 && max > 0) {
-      updateParams.ExpressionAttributeValues[':max'] = max - count;
-      updateParams.ConditionExpression = 'semvalue <= :max';
-    }
-    return this.docClient.update(updateParams).promise();
-  }
-}
-
 class Mutex {
   constructor(docClient, tableName) {
     this.docClient = docClient;
@@ -203,7 +132,6 @@ class Mutex {
 
 module.exports = {
   Mutex: Mutex,
-  Semaphore: Semaphore,
   limit: limit,
   mapTolerant: mapTolerant,
   promiseUrl: promiseUrl,
